backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
jinja2==3.1.4
python-dotenv==1.0.1
alembic==1.13.1


# F:\AI_Projects\LLMChtbot\llmenv\Scripts\activate
# cd doctor-app/backend
# uvicorn main:app --reload
# cd doctor-app/frontend
# npm install
# npm run dev
